requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==3.2.3
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.0
pycparser==2.22
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
